+++ src/system_init.hpp
@@ -1,8 +1,8 @@
-std::vector<std::vector<float> > positions_init(int agent_number, 
-    float box_size, int dim, std::mt19937& gen, std::string init_mode = "random_uniform")
+std::vector<std::vector<double> > positions_init(int agent_number, 
+    double box_size, int dim, std::mt19937& gen, std::string init_mode = "random_uniform")
 {
-    std::vector<std::vector<float> > positions(agent_number, 
-        std::vector<float>(dim));
+    std::vector<std::vector<double> > positions(agent_number, 
+        std::vector<double>(dim));
     // saves current positions for each agent
     
     if (init_mode == "random_uniform")
@@ -23,12 +23,12 @@ std::vector<std::vector<float> > positions_init(int agent_number,
     return positions;
 }
 
-std::vector<std::vector<float> > angles_init(int agent_number, 
-    float box_size, int dim, float angle_interval_low, float angle_interval_high,
+std::vector<std::vector<double> > angles_init(int agent_number, 
+    double box_size, int dim, double angle_interval_low, double angle_interval_high,
     std::mt19937& gen, std::string init_mode = "random_uniform")
 {
-    std::vector<std::vector<float> > angles(agent_number, 
-        std::vector<float>(dim-1));
+    std::vector<std::vector<double> > angles(agent_number, 
+        std::vector<double>(dim-1));
     // saves current angles
  
     if (init_mode == "random_uniform")
